Fixes #21 - When searching for India, you should only get India.
Fixes #21
https://github.com/MobMin/digital_atlas/issues/21

Changed the search criteria so that when searching for a country it matches it only
adds the wildcard at the end of the search parameter.  So searching for "Ind" will
do a LIKE statement "Ind%" finding anything that starts with those letters. Rather then
finding those letters any where in the string.

- We assume user know the first letters for a country.
- Countries are in alphetical naming.